refactor: LineController에서 SectionController 추출
