+++ src/main/java/wooteco/subway/controller/SectionController.java
@@ -1,0 +1,36 @@
+package wooteco.subway.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import wooteco.subway.dto.request.SectionRequest;
+import wooteco.subway.service.SectionService;
+
+@RestController
+@RequestMapping("/lines/{id}/sections")
+public class SectionController {
+
+    private final SectionService sectionService;
+
+    public SectionController(SectionService sectionService) {
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping()
+    public ResponseEntity<Void> createSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
+        sectionService.save(SectionConverter.toInfo(id, sectionRequest));
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping()
+    public ResponseEntity<Void> deleteSection(@PathVariable Long id, @RequestParam Long stationId) {
+        sectionService.delete(SectionConverter.toInfo(id, stationId));
+        return ResponseEntity.ok().build();
+    }
+}
